mt5-framework/requirements.txt
python-dotenv
pydantic
pandas
MetaTrader5
